NextPageDetail finish
