+++ src/components/SVG/settingsSVG/SettingsActivitySvg.jsx
