+++ src/components/SVG/settingsSVG/SettingsFriendsSVG.jsx
