+++ src/components/SVG/settingsSVG/SettingsInfoSvg.jsx
