+++ src/components/SVG/settingsSVG/SettingsQRSvg.jsx
